Change how compressed codes are stored
Before, bits to be used for compression are stored in CodeWord class.
However, the class has a fixed storage size, which makes it unsuitable
when the bits become too long. To fix it, we switch between a
fixed-size buffer and a dynamic one base on the max length of bits.

* Implement BaseBitset and Bitset32 classes
* Implement bit twiddling utility functions in BitUtil namespace
* Remove CodeWord class and replace with BaseBitSet