DynamicBitset: optimize stream output operator
Previously each byte in `DynamicBitset` is retrieved through the
`getByte` interface for writing to an `ostream`. However, since
we're writing byte per byte, this process is pretty inefficient.

Now, we directly write each `Block` (which is usually `uint64_t`)
by casting them to `const char*`. This seems to substantially improve
the performance of stream write operator.